requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
requests==2.32.3
beautifulsoup4==4.12.3
trafilatura==1.9.0
readability-lxml==0.8.1
markdownify==0.11.6
openai==1.37.1
sqlalchemy==2.0.31
psycopg[binary]==3.2.1
pgvector==0.2.5
python-dotenv==1.0.1
tiktoken==0.7.0
pydantic==2.7.4
